# Pre-calculate all non-active buckets in Epsilon Greedy Pool