--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,7 @@
 language: go
 
 go:
-  - 1.13.x
-  - tip
+  - 1.15.x
+  - 1.16.x
 
 script: go test -v -benchmem -run=".*" ./... -bench "Benchmark.*"