--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/bitly/go-hostpool
+
+go 1.15
+
+require github.com/stretchr/testify v1.5.1