--- a/hostpool_test.go
+++ b/hostpool_test.go
@@ -146,33 +146,42 @@ func BenchmarkEpsilonGreedy(b *testing.B) {
 	}
 }
 
-func BenchmarkEpsilonGreedyManyHosts(b *testing.B) {
-	b.StopTimer()
+func BenchmarkEpsilonGreedyManyHosts(topB *testing.B) {
+	bench := func(hostCount int) func(*testing.B) {
+		return func(b *testing.B) {
+			b.StopTimer()
 
-	// Make up some response times
-	zipfDist := rand.NewZipf(rand.New(rand.NewSource(0)), 1.1, 5, 5000)
-	timings := make([]uint64, b.N)
-	for i := 0; i < b.N; i++ {
-		timings[i] = zipfDist.Uint64()
-	}
+			// Make up some response times
+			zipfDist := rand.NewZipf(rand.New(rand.NewSource(0)), 1.1, 5, 5000)
+			timings := make([]uint64, b.N)
+			for i := 0; i < b.N; i++ {
+				timings[i] = zipfDist.Uint64()
+			}
 
-	// Make the hostpool with a few hosts
-	const hostCount = 50
-	hosts := make([]string, 0, hostCount)
-	for i := 0; i < hostCount; i++ {
-		hosts = append(hosts, fmt.Sprintf("%d", i))
-	}
-	p := NewEpsilonGreedy(hosts, 0, &LinearEpsilonValueCalculator{}).(*epsilonGreedyHostPool)
+			// Make the hostpool with a few hosts
+			hosts := make([]string, 0, hostCount)
+			for i := 0; i < hostCount; i++ {
+				hosts = append(hosts, fmt.Sprintf("%d", i))
+			}
+			p := NewEpsilonGreedy(hosts, 0, &LinearEpsilonValueCalculator{}).(*epsilonGreedyHostPool)
 
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		if i != 0 && i%100 == 0 {
-			p.performEpsilonGreedyDecay()
+			b.StartTimer()
+			for i := 0; i < b.N; i++ {
+				if i != 0 && i%100 == 0 {
+					p.performEpsilonGreedyDecay()
+				}
+				hostR := p.Get()
+				p.timer = &mockTimer{t: int(timings[i])}
+				hostR.Mark(nil)
+			}
 		}
-		hostR := p.Get()
-		p.timer = &mockTimer{t: int(timings[i])}
-		hostR.Mark(nil)
 	}
+
+	topB.Run("Hosts10", bench(10))
+	topB.Run("Hosts25", bench(25))
+	topB.Run("Hosts50", bench(50))
+	topB.Run("Hosts100", bench(100))
+	topB.Run("Hosts250", bench(250))
 }
 
 func TestHostPoolErrorBudget(t *testing.T) {