This PR makes some substantial performance improvements to the Epsilon Greedy Hostpool.

Before I explain the performance improvements, I need to explain a little about how the scoring calculation works per host. Let's take an example of a host with 5 stored buckets. The pool keeps track of two core values per bucket, a count and a cumulative duration total (in nanoseconds). 

The goal is to keep track of performance over time but also weigh real time performance higher when scoring.

The list of buckets is a fixed size ring buffer, here we use 5 (but in practice go-hostpool uses 120 buckets). 
- Each bucket has an proportionally reduced weighting depending on how old it is (so older buckets have a reduced weighting, more recent buckets have an increased weighting). 
- The current bucket has a 100% weighting to the score (as it is the most recent values)

![image](https://user-images.githubusercontent.com/76773/109498846-f4bcf880-7a8b-11eb-80ba-78ee5be81d45.png)

So in the example above, the score for this particular host will be calculated as (starting from the oldest bucket)
```
      6 ×  540 × (1 / 5.0)
  +  15 × 1495 × (2 / 5.0)
  +  11 ×  900 × (3 / 5.0)
  +  10 × 1500 × (4 / 5.0)
  +  12 × 1140 × (5 / 5.0)
```

When decay happens, the current bucket and oldest bucket offsets move one index forward, resetting the current bucket to zero.

After some profiling, I found this calculation to be taking a substantial part of CPU time and was being done _every time we requested a host_. I found that this entire calculation could be optimised substantially to pre-calculate most of it. Only the average of the current bucket is changing in this calculation. We can precompute all the other bucket totals and just add the current bucket average on demand

![image](https://user-images.githubusercontent.com/76773/109500075-a7418b00-7a8d-11eb-88f6-ffdb5d7af1ce.png)

This means on each decay, we pre-compute the following sum one time on decay and essentially 'cache' it
```
      6 ×  540 × (1 / 5.0)
  +  15 × 1495 × (2 / 5.0)
  +  11 ×  900 × (3 / 5.0)
  +  10 × 1500 × (4 / 5.0)
```

Then when we request the current score of the host, we take the current bucket average (which is weighted at 100% anyway) and add it on, thus saving four other calculations in this example (and 119 per host in real life).

Before these changes, here's how the performance stacked up
```
pkg: github.com/bitly/go-hostpool
BenchmarkEpsilonGreedy
BenchmarkEpsilonGreedy-8            	  449230	      2820 ns/op	     127 B/op	       3 allocs/op
BenchmarkEpsilonGreedyManyHosts
BenchmarkEpsilonGreedyManyHosts/Hosts10
BenchmarkEpsilonGreedyManyHosts/Hosts10-8         	  105246	     12184 ns/op	     349 B/op	       6 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25
BenchmarkEpsilonGreedyManyHosts/Hosts25-8         	   42470	     29409 ns/op	     601 B/op	       7 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50
BenchmarkEpsilonGreedyManyHosts/Hosts50-8         	   21153	     59977 ns/op	    1070 B/op	       8 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100
BenchmarkEpsilonGreedyManyHosts/Hosts100-8        	   10000	    114046 ns/op	    1663 B/op	       9 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250
BenchmarkEpsilonGreedyManyHosts/Hosts250-8        	    5290	    268156 ns/op	    1047 B/op	       8 allocs/op
```

This meant if you have 100 hosts, you'll be burning 114046 nanoseconds of CPU time (under a mutual lock also) fetching a host. That's a very long amount of time!

With these changes, things get substantially better!
```
pkg: github.com/bitly/go-hostpool
BenchmarkEpsilonGreedy
BenchmarkEpsilonGreedy-8            	 2128981	       559 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts
BenchmarkEpsilonGreedyManyHosts/Hosts10
BenchmarkEpsilonGreedyManyHosts/Hosts10-8         	 1425793	       768 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25
BenchmarkEpsilonGreedyManyHosts/Hosts25-8         	 1000000	      1115 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50
BenchmarkEpsilonGreedyManyHosts/Hosts50-8         	  739261	      1719 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100
BenchmarkEpsilonGreedyManyHosts/Hosts100-8        	  474729	      3141 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250
BenchmarkEpsilonGreedyManyHosts/Hosts250-8        	  246646	      6379 ns/op	     104 B/op	       2 allocs/op
```

That's a 30x to 50x improvement by being a bit more intelligent!