@suhailpatel 2021-03-01
I tweaked the benchmark a little because we're decaying after every 100 ticks which in the much faster version is quite often. The decay in reality occurs once on a timer as mentioned on the original post and not every time we request a host.

Here's the benchmark results with decay disabled and decay enabled.

Before
```
pkg: github.com/bitly/go-hostpool
BenchmarkEpsilonGreedyManyHosts
BenchmarkEpsilonGreedyManyHosts/Hosts10/NoDecay-8         	  109764	     11249 ns/op	     349 B/op	       6 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25/NoDecay-8         	   44950	     27166 ns/op	     602 B/op	       7 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50/NoDecay-8         	   23008	     51590 ns/op	    1084 B/op	       8 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100/NoDecay-8        	   10000	    102160 ns/op	    1757 B/op	       9 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250/NoDecay-8        	    5642	    240300 ns/op	    1428 B/op	       9 allocs/op

BenchmarkEpsilonGreedyManyHosts/Hosts10/WithDecay-8       	  102639	     11637 ns/op	     349 B/op	       6 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25/WithDecay-8       	   44194	     28124 ns/op	     601 B/op	       7 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50/WithDecay-8       	   21825	     57196 ns/op	    1095 B/op	       8 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100/WithDecay-8      	   10000	    111213 ns/op	    1648 B/op	       9 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250/WithDecay-8      	    5392	    247852 ns/op	    1220 B/op	       8 allocs/op
```

After
```
pkg: github.com/bitly/go-hostpool
BenchmarkEpsilonGreedyManyHosts
BenchmarkEpsilonGreedyManyHosts/Hosts10/NoDecay-8         	 2070523	       574 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25/NoDecay-8         	 1719786	       660 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50/NoDecay-8         	 1417652	       841 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100/NoDecay-8        	 1000000	      1280 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250/NoDecay-8        	  584746	      2701 ns/op	     104 B/op	       2 allocs/op

BenchmarkEpsilonGreedyManyHosts/Hosts10/WithDecay-8       	 1878009	       643 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts25/WithDecay-8       	 1238664	       923 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts50/WithDecay-8       	  924391	      1439 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts100/WithDecay-8      	  569004	      2708 ns/op	     104 B/op	       2 allocs/op
BenchmarkEpsilonGreedyManyHosts/Hosts250/WithDecay-8      	  282620	      5691 ns/op	     104 B/op	       2 allocs/op
```

Looking at the decay disabled numbers for comparison, that's an ~85x improvement for the hot path 🤯