@suhailpatel 2021-03-01
I've fixed the use of the calculator, that was some testing on my part, we should use the calculator provided especially because that's part of the exposed interface

> Is the intuition that the active bucket is the only thing which needs to be checked on the 'hot' path, and we can shift the calculation of all other buckets into the background decay step?

Exactly, we keep a running total with everything else in the decay step, then we can just 'append' the calculation for the current bucket to come up with the overall score.